moved utils related commands to their own plugin
This is part 3 of #365 after #401.
